Update networking section of components overview
The section shares future plans on how advanced networking should be
done in KubeVirt. Since it was written, we have advanced in this area
and that should be reflected there.

This patch updates it to reflect the current architecture and focus.

It does not mention any specific plugins or extensions on purpose, so
the section does not become outdated with future integrations or bugs.
We should track details about APIs and extensions in the user-guide
instead and keep this document high-level.

Fixes: https://github.com/kubevirt/kubevirt/issues/7066

Signed-off-by: Petr Horáček <phoracek@redhat.com>